bpf: move applyConstants to new constants.go
