+++ pkg/bpf/constants.go
@@ -1,0 +1,44 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package bpf
+
+import (
+	"fmt"
+
+	"github.com/cilium/ebpf"
+
+	"github.com/cilium/cilium/pkg/datapath/config"
+)
+
+// applyConstants sets the values of BPF C runtime configurables defined using
+// the DECLARE_CONFIG macro.
+func applyConstants(spec *ebpf.CollectionSpec, obj any) error {
+	if obj == nil {
+		return nil
+	}
+
+	constants, err := config.Map(obj)
+	if err != nil {
+		return fmt.Errorf("converting struct to map: %w", err)
+	}
+
+	for name, value := range constants {
+		constName := config.ConstantPrefix + name
+
+		v, ok := spec.Variables[constName]
+		if !ok {
+			return fmt.Errorf("can't set non-existent Variable %s", name)
+		}
+
+		if v.SectionName != config.Section {
+			return fmt.Errorf("can only set Cilium config variables in section %s (got %s:%s), ", config.Section, v.SectionName, name)
+		}
+
+		if err := v.Set(value); err != nil {
+			return fmt.Errorf("setting Variable %s: %w", name, err)
+		}
+	}
+
+	return nil
+}
